+++ apps/web/composables/useRestrictedAddress/useRestrictedAddress.ts
@@ -1,0 +1,12 @@
+export const useRestrictedAddress = () => {
+  const route = useRoute();
+  const localePath = useLocalePath();
+
+  const state = useState('useRestrictedAddress', () => ({
+    restrictedAddresses: route.fullPath.includes(localePath(paths.readonlyCheckout)),
+  }));
+
+  return {
+    ...toRefs(state.value),
+  };
+};
